Rewrite AND Q-expressions to safe form
